implementing the game logic in totality
