+++ src/PlayAgain.js
@@ -1,9 +1,15 @@
 import React from "react";
 
-const PlayAgain = ({ resetGame }) => {
+const PlayAgain = ({ resetGame, gameStatus }) => {
   return (
     <div className='game-done'>
-      <button onClick={resetGame}>Play Again</button>
+      <div
+        className='message'
+        style={{ color: gameStatus === "lost" ? "red" : "green" }}
+      >
+        {gameStatus === "lost" ? "Game over" : "Nice"}
+      </div>
+      <button className='play-again' onClick={resetGame}>Play Again</button>
     </div>
   );
 };
