+++ src/StarMatch.js
@@ -1,81 +1,10 @@
 import React, { useState } from "react";
-import { utils } from "./utils";
-import "./StarMatch.css";
-import PlayNumber from "./PlayNumber";
-import StarsDisplay from "./StarDisplay";
-import PlayAgain from "./PlayAgain";
+import './StarMatch.css';
+import Game from "./Game";
 
 const StarMatch = () => {
-  const [stars, setStars] = useState(utils.random(1, 9));
-  const [candidateNums, setCandidateNums] = useState([]);
-  const [availableNums, setAvailableNums] = useState(utils.range(1, 9));
-
-  const candidatesAreWrong = utils.sum(candidateNums) > stars;
-  const gameIsDone = availableNums.length === 0;
-
-  const resetGame = () => {
-    setStars(utils.random(1, 9));
-    setAvailableNums(utils.range(1, 9));
-    setCandidateNums([]);
-  };
-
-  const numberStatus = (number) => {
-    if (!availableNums.includes(number)) {
-      return "used";
-    }
-    if (candidateNums.includes(number)) {
-      return candidatesAreWrong ? "wrong" : "candidate";
-    }
-    return "available";
-  };
-  const onNumberClick = (number, currentStatus) => {
-    if (currentStatus === "used") {
-      return;
-    }
-    const newCandidateNums =
-      currentStatus === "available"
-        ? candidateNums.concat(number)
-        : candidateNums.filter((cn) => cn !== number);
-    if (utils.sum(newCandidateNums) !== stars) {
-      setCandidateNums(newCandidateNums);
-    } else {
-      const newAvailableNums = availableNums.filter(
-        (n) => !newCandidateNums.includes(n)
-      );
-      setStars(utils.randomSumIn(newAvailableNums, 9));
-      setAvailableNums(newAvailableNums);
-      setCandidateNums([]);
-    }
-  };
-  return (
-    <div className='game'>
-      <div className='help'>
-        Pick 1 or more numbers that sum to the number of stars
-      </div>
-      <div className='body'>
-        <div className='left'>
-          {gameIsDone ? (
-            <PlayAgain resetGame={resetGame} />
-          ) : (
-            utils
-              .range(1, stars)
-              .map((starId) => <StarsDisplay key={starId} className='star' />)
-          )}
-        </div>
-        <div className='right'>
-          {utils.range(1, 9).map((num) => (
-            <PlayNumber
-              key={num}
-              number={num}
-              status={numberStatus(num)}
-              onClick={onNumberClick}
-            />
-          ))}
-        </div>
-      </div>
-      <div className='timer'>Time Remaining: 10</div>
-    </div>
-  );
+  const [gameId, setGameId] = useState(1);
+  return <Game key={gameId} startNewGame={() => setGameId(gameId + 1)} className='StarMatch' />;
 };
 
 export default StarMatch;
